# output: use Hash.new {|hash, key| ... } for default value